Move FakeContext and FakeSharedPreferences to support-test-fakes component
