+++ components/feature/search/build.gradle
@@ -49,6 +49,7 @@ dependencies {
     implementation Dependencies.kotlin_stdlib
 
     testImplementation project(':support-test')
+    testImplementation project(':support-test-fakes')
     testImplementation project(':support-test-libstate')
     testImplementation Dependencies.androidx_test_core
     testImplementation Dependencies.androidx_test_junit
