+++ components/support/ktx/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     implementation Dependencies.kotlin_coroutines
 
     testImplementation project(':support-test')
+    testImplementation project(':support-test-fakes')
 
     androidTestImplementation project(':support-android-test')
     androidTestImplementation Dependencies.androidx_test_core
